Added dynamic allocation for dynamic buffers

+++ Prism-Core/Src/Prism-Core/Render/RenderContext.h
@@ -13,17 +13,17 @@ class TextureView;
 
 struct DrawCommandDesc
 {
-	int32_t numVertices = 0;
-	int32_t numInstances = 1;
-	int32_t startVertexLocation = 0;
+	int64_t numVertices = 0;
+	int64_t numInstances = 1;
+	int64_t startVertexLocation = 0;
 };
 
 struct DrawIndexedCommandDesc
 {
-	int32_t numIndices = 0;
-	int32_t numInstances = 1;
-	int32_t startIndexLocation = 0;
-	int32_t baseVertexLocation = 0;
+	int64_t numIndices = 0;
+	int64_t numInstances = 1;
+	int64_t startIndexLocation = 0;
+	int64_t baseVertexLocation = 0;
 };
 
 enum class IndexBufferFormat
@@ -52,7 +52,7 @@ public:
 	void SetScissor(Scissor scissor);
 	virtual void SetScissors(std::vector<Scissor> scissors) = 0;
 
-	virtual void SetVertexBuffer(Buffer* buffer, int32_t vertexSizeInBytes) = 0;
+	virtual void SetVertexBuffer(Buffer* buffer, int64_t vertexSizeInBytes) = 0;
 	virtual void SetIndexBuffer(Buffer* buffer, IndexBufferFormat format) = 0;
 
 	virtual void SetTexture(TextureView* textureView, const std::wstring& paramName) = 0;
@@ -68,7 +68,7 @@ public:
 	virtual void CopyBufferRegion(Buffer* dest, int32_t destOffset, Buffer* src, int32_t srcOffset, int32_t numBytes) = 0;
 
 
-	// Object MUST be std::move'd into this function
+	// Objects MUST be std::move'd into this function
 	template<typename T>
 	void SafeReleaseResource(T&& resource) requires !std::is_lvalue_reference_v<T>
 	{
